added logic to get base64 image string, added update image function, changed the models a bit, configured the mockup worker (EVERYTHING CORRELATED TO MOCKUP GENERATION WORKS)

+++ server/app.py
@@ -2,17 +2,20 @@ from flask import Flask, jsonify, request
 import pika
 import pika.exceptions
 from models import db, Image, Mockup
+import json
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
 
 app = Flask(__name__)
+app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY')
+app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('SQLALCHEMY_DATABASE_URI')
+db.init_app(app)
 
 @app.route('/')
 def index():
-    return 'index'
-
-# TODO
-# fill out these methods
-# Pass image prompt to rabbitmq task
-# create models
+    return 'index', 200
 
 # Get image from the database by id
 @app.route('/get-image/<image_id>', methods=['GET'])
@@ -21,10 +24,10 @@ def get_image(image_id):
 
     if image:
         data = image.serialize()
-        return 200, jsonify(data)
+        return jsonify(data), 200
     
     else:
-        return 404, jsonify({'error': 'Image with such id was not found'})
+        return jsonify({'error': 'Image with such id was not found'}), 404
 
 # Delete image from the database by id
 @app.route('/delete-image/<image_id>', methods=['POST'])
@@ -32,12 +35,35 @@ def delete_image(image_id):
     if request.method == 'POST':
         image = Image.query.filter_by(id=image_id).first()
 
-        db.session.delete(image)
-        db.session.commit()
+        if image:
+            db.session.delete(image)
+            db.session.commit()
 
-        return 200, 'Image deleted successfully'
+            return 'Image deleted successfully', 200
+        
+        else:
+            return jsonify({'error': 'Image with such id was not found'}), 404
+    
     else:
-        return 400, jsonify({'error': 'Invalid request. Use POST request'})
+        return jsonify({'error': 'Invalid request. Use POST request'}), 400
+
+@app.route('/update-image/<image_id>', methods=['POST'])
+def update_image(image_id):
+    if request.method == 'POST':
+        data = request.get_json(force=True)
+        image = Image.query.filter_by(id=image_id).first()
+
+        if image:
+            image.url = data['ai_image_url']
+            db.session.commit()
+
+            return 'Image updated successfully', 200
+        
+        else:
+            return jsonify({'error': 'Image with such id was not found'}), 404
+    
+    else:
+        return jsonify({'error': 'Invalid request. Use POST request'}), 400
 
 # Create a mockup
 @app.route('/create-mockup/', methods=['POST'])
@@ -45,14 +71,14 @@ def create_mockup():
     if request.method == 'POST' and request.data:
         data = request.get_json(force=True)
 
-        mockup = Mockup(data['title'], data['price'], data['color'], data['size'], data['mockup_image_url'], data['ai_image_url'], data['printful_product_id'], data['printful_variant_id'])
+        mockup = Mockup(data['title'], data['price'], data['color'], data['mockup_image_url'], data['ai_image_id'], data['printful_product_id'])
         db.session.add(mockup)
         db.session.commit()
     
-        return 201, 'Mockup was successfully created'
+        return 'Mockup was successfully created', 201
     
     else:
-        return 400, jsonify({'error': 'There was an error in creating the mockup'})
+        return jsonify({'error': 'There was an error in creating the mockup'}), 400
 
 # Get mockup from database by id
 @app.route('/get-mockup/<mockup_id>', methods=['GET'])
@@ -61,10 +87,10 @@ def get_mockup(mockup_id):
 
     if mockup:
         data = mockup.serialize()
-        return 200, jsonify(data)
+        return jsonify(data), 200
     
     else:
-        return 404, jsonify({'error': 'Mockup with such id was not found'})
+        return jsonify({'error': 'Mockup with such id was not found'}), 404
 
 # Delete mockup from database by id
 @app.route('/delete-mockup/<mockup_id>', methods=['POST'])
@@ -75,42 +101,51 @@ def delete_mockup(mockup_id):
         db.session.delete(mockup)
         db.session.commit()
 
-        return 'Mockup deleted successfully'
+        return 'Mockup deleted successfully', 200
     else:
-        return 400, jsonify({'error': 'Invalid request. Use POST request'})
+        return jsonify({'error': 'Invalid request. Use POST request'}), 404
 
 # Creates a mockup generation task with RabbitMQ and returns the image id to which mockups will be generated
-@app.route('/image-generator/create-task/', methods=['POST'])
+@app.route('/mockup-generator/create-task/', methods=['POST'])
 def create_mockup_task():
     if request.method == 'POST':
         data = request.get_json(force=True)
-        prompt = data['prompt']
 
-        image = Image(prompt, None)
-
+        image = Image(data['prompt'], None)
         db.session.add(image)
         db.session.commit()
 
+        body = {
+            'prompt': data['prompt'],
+            'ai_image_id': image.id
+        }
+
         try:
-            connection = pika.BlockingConnection(pika.ConnectionParameters(host='localhost:5672'))
+            connection = pika.BlockingConnection(pika.ConnectionParameters(host='localhost', port=5672))
         except pika.exceptions.AMQPConnectionError:
-            return 500, jsonify({'error': 'Failed to connect to RabbitMQ service. Message wont be sent.'})
+            return jsonify({'error': 'Failed to connect to RabbitMQ service. Message wont be sent.'}), 500
         
         channel = connection.channel()
-        channel.queue_declare('image_generation_queue', durable=True)
+        channel.queue_declare('mockup_generation_queue', durable=True)
 
         channel.basic_publish(
             exchange = '',
-            routing_key = 'image_generation_queue',
+            routing_key = 'mockup_generation_queue',
             properties = pika.BasicProperties(
                 delivery_mode = pika.DeliveryMode.Persistent
             ),
-            body=prompt
+            body=json.dumps(body)
         )
 
         connection.close()
 
-        return 200, jsonify({'image_id': image.id})
+        return jsonify({'image_id': image.id}), 200
     
     else:
-        return 400, jsonify({'error': 'Invalid request. Use POST request'})
+        return jsonify({'error': 'Invalid request. Use POST request'}), 400
+    
+if __name__ == '__main__':
+    with app.app_context():
+        db.create_all()
+        db.session.commit()
+    app.run(debug=True)
